LayoutTests/fast/selectors/querySelector-leftmost-selector-matches-rootNode.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="target" class="target">
  <div></div>
  <div></div>
  <div></div>
  <div></div>
  <div></div>
  <div></div>
  <pre id='console'></pre>
</body>
<script>
description('Test queries whose leftmost selector matches the root node.');
shouldBe("document.body.querySelectorAll('#target > div').length", "6");
shouldBe("document.body.querySelectorAll('#target').length", "0");
shouldBe("document.body.querySelectorAll('.target > div').length", "6");
shouldBe("document.body.querySelectorAll('.target').length", "0");
shouldBe("document.body.querySelectorAll('body > div').length", "6");
shouldBe("document.body.querySelectorAll('body').length", "0");
shouldBeNull("document.body.querySelector('#target')");
shouldBeNull("document.body.querySelector('.target')");
shouldBeNull("document.body.querySelector('body')");
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
